feat: 위치는 0보다 커야한다
