+++ README.md
@@ -3,6 +3,8 @@
 ## TDD 기능 요구사항
 ### 지뢰 찾기를 변형한 프로그램을 구현한다.
 -[x] 높이와 넓이 갯수는 숫자이어야한다.
+-[x] 위치는 0보다 커야한다
+-[ ] 위치는 입력받은 높이와 넓이의 갯수보다 작아야한다
 -[ ] 지뢰는 눈에 잘 띄는 것으로 표기한다.
 -[ ] 지뢰는 가급적 랜덤에 가깝게 배치한다.
 -[ ] 각 사각형에 표시될 숫자는 자신을 제외한 주변 8개 사각형에 포함된 지뢰의 개수다.
