+++ src/main/kotlin/minesweeper/MineSweeperGameMain.kt
@@ -1,0 +1,5 @@
+package minesweeper
+
+fun main() {
+
+}
