+++ src/main/kotlin/minesweeper/domain/MinePosition.kt
@@ -1,5 +1,15 @@
 package minesweeper.domain
 
-data class MinePosition(val positionX: Position, val positionY: Position) {
-    constructor(positions: List<Position>) : this(positions[0], positions[1])
+
+class MinePosition private constructor(val x: Int, val y: Int) {
+    companion object {
+        fun of(positionX: Int, positionY: Int): MinePosition {
+
+            require(positionX > 0 && positionY > 0) {
+                "위치값은 0보다 커야한다"
+            }
+            return MinePosition(positionX, positionY)
+        }
+    }
+
 }
