+++ src/main/kotlin/minesweeper2/domain/MineTile.kt
@@ -1,4 +1,4 @@
-package minesweeper.domain
+package minesweeper2.domain
 
 class MineTile : Tile() {
     override fun getType(): TileType {
