+++ src/main/kotlin/minesweeper2/domain/Position.kt
@@ -1,4 +1,4 @@
-package minesweeper.domain
+package minesweeper2.domain
 
 @JvmInline
 value class Position(val position: Int) {
