+++ src/main/kotlin/minesweeper2/domain/RandomPosition.kt
@@ -1,4 +1,4 @@
-package minesweeper.domain
+package minesweeper2.domain
 
 import kotlin.random.Random
