--- src/test/kotlin/minesweeper2/DimensionTest.kt
@@ -1,18 +1,0 @@
-package minesweeper2
-
-import io.kotest.assertions.throwables.shouldThrow
-import io.kotest.core.spec.style.StringSpec
-import minesweeper.Dimension
-
-class DimensionTest : StringSpec({
-    "높이와 넓이 갯수는 숫자이어야한다" {
-        shouldThrow<IllegalArgumentException> {
-            Dimension.of("a")
-        }
-    }
-    "높이와 넓이 갯수는 0보다 커야한다" {
-        shouldThrow<IllegalArgumentException> {
-            Dimension.of("0")
-        }
-    }
-})
